Introduce `sgx.remote_attestation = "[none|epid|dcap]"`
This commit deprecates ambiguous `sgx.remote_attestation = true` in
favor of an explicit attestation-type string. The first side benefit of
this change is that the weird logic "if ra_client_spid is set, then it
is EPID otherwise DCAP" is gone. The second side benefit is that this
allows us to add more SGX quote formats in the future, if needed.

Also, a new pseudo-file `/dev/attestation/attestation_type` is
introduced at the LibOS level. It helps the application to determine
which attestation scheme should be used.

Several CI examples are rewritten to use the new syntax: Python,
ra-tls-mbedlts, ra-tls-secret-prov, Busybox (the latter is purely for
testing purposes). The `attestation` LibOS regression test also uses the
new syntax; the old syntax is tested via the newly added test
`attestation_deprecated_syntax`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>